first commit of new java framework
This is the first commit of a new framework i’m working on that is to
support dynamic creation of html using an EL and enum’s also with the
backing of a dynamic model creation which sits on mongodb.  This is the
idea that using the play framework and annotations as well as JIT to
create a dynamic web model that can adapt to user input while
maintaining security and robustness.  Pretty much I’m working on a high
level web development language which will allow for rapped changes to
models html etc… with little change on the user side.